Complete Project

+++ BE/src/main/java/org/kltn/postconnector/api/repository/OrderItemRepository.java
@@ -1,13 +1,12 @@
 package org.kltn.postconnector.api.repository;
 
-import org.kltn.postconnector.api.model.Order;
-import org.kltn.postconnector.api.model.OrderItem;
+import org.kltn.postconnector.api.domain.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
-    void deleteAllByOrderId(long orderId);
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+    void deleteAllByOrderId(int orderId);
 
-    List<OrderItem> findAllByOrderId(long orderId);
+    List<OrderItem> findAllByOrderId(int orderId);
 }
